project/llm_sql_query_system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.4.2
python-dotenv==1.0.0
PyYAML==6.0.1
psycopg2-binary==2.9.9
openai==1.3.5